bumped ark version used to 0.4.2, mibi-bin-tools version to 0.2.4, and toffy version to 0.1.1

+++ toffy/normalize.py
@@ -714,7 +714,7 @@ def normalize_image_data(img_dir, norm_dir, pulse_height_dir, panel_info,
 
         # get images
         images = load_utils.load_imgs_from_tree(img_dir, fovs=[fov], channels=channels,
-                                                dtype='float32', img_sub_folder=img_sub_folder)
+                                                img_sub_folder=img_sub_folder)
 
         # normalize and save
         normalize_fov(img_data=images, norm_vals=norm_vals, norm_dir=norm_dir, fov=fov,
